requirements.txt
Django==2.2.3
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
pycryptodome==3.4.3
Pillow==6.1.0
Unidecode==1.1.1
django-phonenumber-field==3.0.1
django-parsley==0.7

